requirements.txt
torch>=2.0.0
transformers>=4.30.0
flask>=2.0.0
sqlalchemy>=1.4.0
speechrecognition>=3.8.1
nltk>=3.6.0
pandas>=1.3.0
scikit-learn>=0.24.0
python-dotenv>=0.19.0
cx_Freeze>=6.11.0
pillow>=8.0.0 